+++ R/env.R
@@ -1076,6 +1076,14 @@ ns_env_name <- function(pkg = NULL) {
   unname(getNamespaceName(pkg))
 }
 
+#' Is an object a namespace environment?
+#'
+#' @param x An object to test.
+#' @export
+is_namespace <- function(x) {
+  isNamespace(x)
+}
+
 #' Is a package installed in the library?
 #'
 #' This checks that a package is installed with minimal side effects.